Rename docker::run to docker::container to prepare for ensure parameter
