+++ manifests/container_instance.pp
@@ -1,0 +1,4 @@
+# docker::container_instance
+class docker::container_instance($instance) {
+  create_resources(docker::container, $instance)
+}
