+++ manifests/init.pp
@@ -228,7 +228,7 @@ class docker(
 
   # Only bother trying extra docker stuff after docker has been installed,
   # and is running.
-  Class['docker'] -> Docker::Registry <||> -> Docker::Run <||>
+  Class['docker'] -> Docker::Registry <||> -> Docker::Container <||>
   Class['docker'] -> Docker::Image <||>
 
 }
