+++ manifests/run_instance.pp
@@ -1,4 +1,4 @@
 # docker::run_instance
 class docker::run_instance($instance) {
-  create_resources(docker::run, $instance)
+  create_resources(docker::container, $instance)
 }
