+++ templates/etc/init.d/docker-container.erb
