+++ templates/etc/systemd/system/docker-container.erb
